apps/api/pyproject.toml
[project]
name = "orka-api"
version = "0.1.0"
description = "orka-api"
requires-python = ">=3.12.12"
dependencies = [
    "asyncpg>=0.31.0",
    "hypercorn>=0.18.0",
    "orka-db",
    "PyJWT>=2.10.1",
    "pydantic>=2.11.10",
    "python-socketio>=5.16.1",
    "sqlalchemy>=2.0.49",
    "starlette>=1.0.0",
]

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"

[tool.uv.sources]
orka-db = { workspace = true }
